Creates a sparse matrix
This was done for my systems programming class as an exercise to get us comfortable with the C programming language. The accompanying header file was written by my instructor though I wrote the main program. It creates a sparse matrix with less storage than a 2x2 array by not storing 0 value nodes.