Create SetFunctions.rkt
This program was the final racket program I made for class. It looks at different sets and does things like finding the intersection, finding the union, determining if they are equivalent, and more. It was intended to teach us the functional programming paradigm so it relies on recursion and does not modify the program state.